+++ examples/authors/csharp/adonet/Models.cs
@@ -42,3 +42,15 @@ public sealed record UpsertAuthorParams(
     string AuthorName,
     string? Biography
 );
+
+public sealed record FindAuthorsByNameRow(
+    ulong ID,
+    string Name,
+    string? Bio
+);
+
+public sealed record FindAuthorsByNameCoveringRow(
+    ulong ID,
+    string Name,
+    string? Bio
+);